Refactor token and AST definitions into their own files
