+++ src/lexer.h
@@ -5,53 +5,12 @@
 #ifndef GRAM_LEXER_H
 #define GRAM_LEXER_H
 
+#include "tokens.h"
+#include <memory>
 #include <string>
 #include <vector>
-#include <memory>
 
 namespace gram {
-
-  enum class TokenType {
-    BEGIN,
-    COLON,
-    END,
-    EQUALS,
-    IDENTIFIER,
-    INTEGER,
-    SEQUENCER,
-    THICK_ARROW,
-    THIN_ARROW
-  };
-
-  const char * const TokenTypeName[] = {
-    "BEGIN",
-    "COLON",
-    "END",
-    "EQUALS",
-    "IDENTIFIER",
-    "INTEGER",
-    "SEQUENCER",
-    "THICK_ARROW",
-    "THIN_ARROW"
-  };
-
-  class Token {
-  public:
-    gram::TokenType type;
-    std::string literal;
-    std::shared_ptr<std::string> source_name;
-    std::shared_ptr<std::string> source;
-    size_t start_pos; // Inclusive
-    size_t end_pos; // Exclusive
-
-    Token(
-      gram::TokenType type, const std::string &literal,
-      std::shared_ptr<std::string> source_name,
-      std::shared_ptr<std::string> source,
-      size_t start_pos, size_t end_pos
-    );
-    std::string show();
-  };
 
   // Perform lexical analysis.
   // The lexer guarantees that all BEGIN/END tokens will be matched
